+++ ChangeLog.md
@@ -37,6 +37,10 @@
     Geodesics in Heat approach and the new differential operators on
     polygonal surfaces (digital surfaces, or any PolygonalMesh instance) (David
     Coeurjolly, [#1603](https://github.com/DGtal-team/DGtal/pull/1603)
+  - Updates to PolygonalCalculus: changing sign convention, fix some Eigen
+    problems, add Dirichlet boundary conditions, update discrete
+    differential calculus examples (Jacques-Olivier
+    Lachaud,[#1643](https://github.com/DGtal-team/DGtal/pull/1643))
 
 - *Topology*
   - New helper methods to retrieve the interior/exterior voxel of a given